source/IoTEdge/EdgeInspect/modules/ImageCaptureModule/requirements.txt
azure-iot-device
requests
opencv-python
faker 
matplotlib
azure-storage-blob
pyodbc
paho-mqtt